+++ packages/devtools_app/lib/src/screens/memory/framework/connected/memory_controller.dart
@@ -11,8 +11,8 @@ import '../../../../shared/memory/class_name.dart';
 import '../../../../shared/memory/heap_graph_loader.dart';
 import '../../panes/chart/controller/android_chart_controller.dart';
 import '../../panes/chart/controller/chart_pane_controller.dart';
+import '../../panes/chart/controller/event_chart_controller.dart';
 import '../../panes/chart/controller/vm_chart_controller.dart';
-import '../../panes/chart/widgets/memory_events_pane.dart';
 import '../../panes/control/controller/control_pane_controller.dart';
 import '../../panes/diff/controller/diff_pane_controller.dart';
 import '../../panes/profile/profile_pane_controller.dart';
